packaging: log rotation - do not execute ls on non-existing files
Due to the usage of ls, ovirtlogrot.sh would print error messages if it
couldn't find any file matching the pattern engine.log.*
This patch removes the execution of ls, and therefore, removes the error
messages.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=889751

+++ packaging/resources/ovirtlogrot.sh
@@ -11,9 +11,9 @@ dateformat="+%Y%m%d_%H%M%S"
 
 date=`date $dateformat`
 
-for i in `ls $logdir/{engine,server,jasperserver}.log.* -t`; do
+for i in $logdir/{engine,server,jasperserver}.log.*; do
  m=`expr match $i '.*\(gz\|xz\)'`
- if [ -z "$m" ]
+ if [ -z "$m" -a -r "$i" ]
  then
   cat $i | xz > $i-${date}_`/bin/date +%N | cut -c6-`.xz && rm $i
  fi
@@ -24,7 +24,7 @@ done;
 #
 
 lastlogday=`date -d "$maxage hours ago" $dateformat`
-for i in `ls $logdir/{engine,server,jasperserver}.log.*.{gz,xz}`; do
+for i in $logdir/{engine,server,jasperserver}.log.*.{gz,xz}; do
  timestamp=`echo $i | sed 's/.*-//; s/\.\(gz\|xz\)$//; s/_.\{4\}$//'`
  if [[ "$lastlogday" > "$timestamp" ]]
  then
